requirements.txt
certifi==2019.3.9
chardet==3.0.4
gunicorn==19.9.0
idna==2.8
libmagic==1.0
python-dotenv==0.10.1
python-magic-bin==0.4.14
requests==2.21.0
six==1.12.0
urllib3==1.24.2
websocket-client==0.56.0
